requirements.txt
faiss-cpu
sentence-transformers
fastapi
uvicorn
pypdf
streamlit
requests
